[WIP] Supporting legacy drivers using the V1 namespace
