+++ lib/browse_everything/browser.rb
@@ -16,7 +16,7 @@ module BrowseEverything
       @providers = ActiveSupport::HashWithIndifferentAccess.new
       opts.each_pair do |driver_key, config|
         driver = driver_key.to_s
-        driver_klass = BrowseEverything::Driver.const_get((config[:driver] || driver).camelize.to_sym)
+        driver_klass = BrowseEverything::V1::Driver.const_get((config[:driver] || driver).camelize.to_sym)
         @providers[driver_key] = driver_klass.new(config.merge(url_options: url_options))
       rescue NameError
         Rails.logger.warn "Unknown provider: #{driver}"
